Move code to main page folder, including tests
