+++ src/dashboard.ts
@@ -1,13 +1,12 @@
 import { readFileSync, existsSync, unlinkSync, rmSync } from 'node:fs';
 import { resolve } from 'node:path';
 import { execFile, ChildProcessPromise } from 'promisify-child-process';
-import { QueryConfig } from 'pg';
 
 import { TimedCacheValidity, Database, DatabaseConfig, Source } from './db.js';
 import { startRequest, completeRequest } from './perf-tracker.js';
-import { AllResults, BenchmarkCompletion } from './api.js';
+import { BenchmarkCompletion } from './api.js';
 import { GitHub } from './github.js';
-import { robustPath, siteConfig, TotalCriterion } from './util.js';
+import { robustPath, siteConfig } from './util.js';
 import { assert, log } from './logging.js';
 import {
   CompareView,
@@ -25,64 +24,6 @@ JOIN Experiment exp ON exp.sourceId = s.id
 JOIN Measurement m ON  m.expId = exp.id
 WHERE repoURL = 'https://github.com/smarr/ReBenchDB'
  */
-
-const resultsCache: AllResults[][] = [];
-let resultsCacheValid: TimedCacheValidity | null = null;
-
-export async function dashResults(
-  projectId: number,
-  db: Database
-): Promise<AllResults[]> {
-  if (
-    resultsCache[projectId] &&
-    resultsCacheValid !== null &&
-    resultsCacheValid.isValid()
-  ) {
-    return resultsCache[projectId];
-  }
-
-  if (resultsCacheValid === null || !resultsCacheValid.isValid()) {
-    resultsCache.length = 0;
-  }
-  resultsCacheValid = db.getStatsCacheValidity();
-
-  const q: QueryConfig = {
-    name: 'all-results',
-    text: ` WITH Results AS (
-              SELECT
-                    value, b.name as benchmark,
-                    rank() OVER (
-                      PARTITION BY b.id
-                      ORDER BY
-                        t.startTime DESC,
-                        m.invocation DESC,
-                        m.iteration DESC
-                    )
-                    FROM Measurement m
-                      JOIN Trial t ON  m.trialId = t.id
-                      JOIN Experiment e ON t.expId = e.id
-                      JOIN Run r ON m.runId = r.id
-                      JOIN Benchmark b ON r.benchmarkId = b.id
-                      JOIN Criterion c ON m.criterion = c.id
-                    WHERE projectId = $1 AND
-                      c.name = '${TotalCriterion}'
-                    ORDER BY t.startTime, m.invocation, m.iteration
-            ),
-            LastHundred AS (
-              SELECT rank, value, benchmark
-              FROM Results
-              WHERE rank <= 100
-              ORDER BY benchmark, rank DESC
-            )
-            SELECT array_agg(value) as values, benchmark
-            FROM LastHundred
-            GROUP BY benchmark;`,
-    values: [projectId]
-  };
-  const result = await db.query(q);
-  resultsCache[projectId] = result.rows;
-  return resultsCache[projectId];
-}
 
 export async function dashProfile(
   runId: number,
